rptest: Move wasm fan in verification function out
- So that other tests may be able to invoke it
